Add an interface for McpServerSession to allow for more extensibility
Having an interface rather than a concrete class
can help with extensibility of the McpServerSession class.
This PR implements that and changes usages around the code
to only mention the interface rather than the implementation
class
